Move fit computation code to dedicated new file (#10174)
* Move fit computation code to dedicated new file

Part of https://github.com/pydata/xarray/issues/10089

* .

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------
